tests/python/getting-started/requirements.txt
# Galileo SDK for OpenAI integration
galileo>=0.0.7

# OpenAI API
openai==1.61.1

# Environment variable management
python-dotenv>=0.19.0

# HTTP requests for API calls
requests>=2.28.0

# Rich terminal formatting
rich>=12.0.0
